분류해 둔 정적 분석 지적을 실제로 고친다 (110 -> 65건)
1회차에서 분류만 해 둔 것을 하나씩 판단해 처리했다. 고친 것은 전부 실물로 확인한 뒤 고쳤고,
오탐은 자리마다 이유를 남겼다.

가장 위험했던 것 — StringUtil 이 비-ASCII 바이트를 부호 있게 다뤘다. char 의 부호성은 구현
정의이고 UTF-8 의 0x80 이상 바이트는 signed char 에서 음수다. 그대로 넓히면서 둘이 깨졌다:

- compare 가 한글처럼 비-ASCII 가 섞인 문자열을 ASCII 보다 작다고 답했다(strcmp 규약의 반대다).
  같은 함수의 대소문자 구분 경로는 이미 uint8 로 비교하고 있어서 두 모드가 서로 다른 순서를 냈다 —
  정렬·이진 검색에 쓰면 일관성이 깨진다.
- computeHash64 는 bIgnoreCase 경로만 부호 확장됐다. 그 인자는 기본값이 true 라 거의 모든 호출이
  이 경로다. 같은 바이트가 경로에 따라 다르게 해싱되고, char 가 unsigned 인 플랫폼에서는 해시
  자체가 달라진다.

둘 다 uint8 을 거치게 했다. ASCII 는 동작이 그대로이고 바뀌는 것은 비-ASCII 뿐이다 — 저장되는
경로·타입 이름은 소문자 ASCII 로 강제돼 있어 안전하다. Core_String.NonAsciiBytesAreUnsigned 로
고정했다. (같은 텍스트라서 utf16 오버로드까지 치환됐던 것을 되돌렸다 — UTF-16 코드 단위를 uint8 로
자르면 값이 잘린다.)

널 역참조 둘:

- ImGuiOpenGLRendererBackend::initialize 가 널·비-OpenGL 을 걸러 _pRHIDevice 를 nullptr 로 만들어
  두고도 Windows 분기에서 검증되지 않은 원시 매개변수를 역참조했다. 바로 아래 Linux 분기는
  막고 있었으니 Windows 쪽만 빠진 것이다. 양쪽을 검증된 멤버로 맞췄다.
- HashedStringPool::shutdown 이 SW_ASSERT 만 믿고 역참조했다. Shipping 에서 단정은 사라지므로
  initialize 전에 또는 두 번 불리면 죽는다.

int 곱셈 후 확대 — Win32SplashWindow 의 픽셀 수·바이트 오프셋(포인터 오프셋으로 쓰인다),
TileMapPanel 의 _width*_height·indexOf, ProfilerPanel 의 바이트 경계, D3D12 디스크립터 오프셋 8곳.
EventDispatcher 는 같은 값의 상수가 이미 있어 그것을 쓴다.

float3/float4x4 에 data() 와 static_assert 를 추가했다. &m._11 을 받아 [0..15] 로 읽는 코드가
여럿 있었다 — 형식적으로 배열이 아닌 멤버를 배열로 읽는 것이라 UB 다. data() 로 모으면 가정이
한 곳에 있고 static_assert 가 패딩 없음을 컴파일 타임에 지킨다. 호출부 4곳을 옮겼다.

ContentBrowser 의 체커보드를 정수 루프로 바꿨다. float 루프 카운터는 반복마다 오차가 쌓이고
칸 색을 정하려고 다시 나눗셈을 해야 했다.

BlendSpace2D 의 가중치 배열을 0 으로 시작한다. empty 를 걸러 실제로는 쓰기 전에 읽히지 않지만
그 사실이 MathUtil::min 을 거친 sampleCount 에 숨어 있다.

ASan 설정을 두 번 고쳤다 — 측정해 보고 판단이 바뀌었다:

- detect_odr_violation 을 1 -> 0 으로. 1 은 "크기가 다른 중복만 보고" 하니 진짜 ODR 버그는 남는다는
  계산이었는데, 레벨 1 도 등록된 전역 전체를 훑는 비용은 그대로 치른다. SmokeTest 가 비-ASan
  6.9초 -> ASan 1800초 초과였고 0 으로 바꾸자 5초였다. 260배는 오버헤드가 아니라 사용 불가다.
- 타임아웃 배수를 sw_addTestExecutable 안에만 두었더니 손으로 add_test 한 EngineTest_NoGPU 가
  빠졌다. 보정을 sw_applySanitizerTestProperties() 한 곳으로 빼고 양쪽이 부른다.
- SmokeTest 는 케이스 하나가 아니라 LiveReload* 전반이 걸리므로 ASan 에서 스위트째 Disabled 다.

결과: ASan 스위트가 1820초/1건 실패 -> 21초/전부 통과.

검증: Debug·Shipping·ASan 빌드 경고 0, lint 6/6, nogpu 5/5 양쪽, EngineTest 423/423,
CoreTest 168/168, 네 백엔드 × 에디터 [Error] 0건, 패널 덤프 14창/빈 0개.

+++ Source/Core/String/StringUtil.cpp
@@ -648,6 +648,9 @@ namespace sw
         return std::wcscmp( pLhs, pRhs ) == 0;
     }
 
+    // 바이트를 **부호 없이** 비교한다. `char` 를 그대로 int 로 넓히면 UTF-8 의 0x80 이상 바이트가
+    // 음수가 되어, 한글처럼 비-ASCII 가 섞인 문자열이 ASCII 보다 **작다고** 나왔다(strcmp 규약의
+    // 반대다). 정렬·이진 검색에 이 함수를 쓰는 쪽이 일관되지 않게 동작한다.
     int32 StringUtil::compare( string_view lhs, string_view rhs, bool bIgnoreCase ) noexcept
     {
         const size_t minLen = ( lhs.size() < rhs.size() ) ? lhs.size() : rhs.size();
@@ -655,8 +658,8 @@ namespace sw
         {
             for ( size_t charIndex = 0; charIndex < minLen; ++charIndex )
             {
-                const int32 c1 = static_cast<int32>( toLowerChar( lhs[charIndex] ) );
-                const int32 c2 = static_cast<int32>( toLowerChar( rhs[charIndex] ) );
+                const int32 c1 = static_cast<int32>( static_cast<uint8>( toLowerChar( lhs[charIndex] ) ) );
+                const int32 c2 = static_cast<int32>( static_cast<uint8>( toLowerChar( rhs[charIndex] ) ) );
                 if ( c1 != c2 )
                     return c1 - c2;
             }
@@ -720,8 +723,8 @@ namespace sw
         {
             while ( *pLhs != '\0' && *pRhs != '\0' )
             {
-                const int32 c1 = static_cast<int32>( toLowerChar( *pLhs ) );
-                const int32 c2 = static_cast<int32>( toLowerChar( *pRhs ) );
+                const int32 c1 = static_cast<int32>( static_cast<uint8>( toLowerChar( *pLhs ) ) );
+                const int32 c2 = static_cast<int32>( static_cast<uint8>( toLowerChar( *pRhs ) ) );
                 if ( c1 != c2 )
                     return c1 - c2;
                 ++pLhs;
